style: add trailing comma to data class, rename App class
