+++ app/src/main/java/kr/ksw/visitkorea/data/remote/dto/AreaCodeDTO.kt
@@ -2,5 +2,5 @@ package kr.ksw.visitkorea.data.remote.dto
 
 data class AreaCodeDTO(
     val code: String,
-    val name: String
+    val name: String,
 )
