+++ app/src/main/java/kr/ksw/visitkorea/data/remote/model/ApiResponse.kt
@@ -1,5 +1,5 @@
 package kr.ksw.visitkorea.data.remote.model
 
 data class ApiResponse<T>(
-    val response: CommonResponse<T>
+    val response: CommonResponse<T>,
 )
