+++ app/src/main/java/kr/ksw/visitkorea/data/remote/model/CommonBody.kt
@@ -4,5 +4,5 @@ data class CommonBody<T>(
     val numOfRows: Int,
     val pageNo: Int,
     val totalCount: Int,
-    val items: CommonItem<T>
+    val items: CommonItem<T>,
 )
