+++ app/src/main/java/kr/ksw/visitkorea/data/remote/model/CommonHeader.kt
@@ -2,5 +2,5 @@ package kr.ksw.visitkorea.data.remote.model
 
 data class CommonHeader(
     val resultCode: String,
-    val resultMsg: String
+    val resultMsg: String,
 )
