+++ app/src/main/java/kr/ksw/visitkorea/data/remote/model/CommonItem.kt
@@ -1,5 +1,5 @@
 package kr.ksw.visitkorea.data.remote.model
 
 data class CommonItem<T>(
-    val item: List<T>
+    val item: List<T>,
 )
