+++ app/src/main/java/kr/ksw/visitkorea/data/remote/model/CommonResponse.kt
@@ -2,5 +2,5 @@ package kr.ksw.visitkorea.data.remote.model
 
 data class CommonResponse<T>(
     val header: CommonHeader,
-    val body: CommonBody<T>
+    val body: CommonBody<T>,
 )
